@@ -1,7 +1,10 @@
-﻿using DuplicateDetector.Services;
+﻿using System.IO;
+using Discord;
+using DuplicateDetector.Services;
 using DuplicateDetector.Templates;
 using Microsoft.Extensions.DependencyInjection;
 using UNObot.Plugins;
+using UNObot.Plugins.Helpers;
 
 namespace DuplicateDetector.Modules
 {
@@ -13,24 +16,45 @@ namespace DuplicateDetector.Modules
         public string Version { get; private set; }
         public IServiceCollection Services { get; private set; }
 
-        public int OnLoad()
+        public int OnLoad(ILogger logger)
         {
+            var configSuccess = LoadConfig(logger, out var config);
+
             Name = "Duplicate Detector";
             Description = "A test program to detect duplicate images. For private use only.";
             Author = "DoggySazHi";
-            Version = "0.1.1 (4.2.0)";
+            Version = config != null ? config.Version : "Unknown Version";
+
+            if (!configSuccess || config == null)
+                return 1;
 
             Services = new ServiceCollection()
-                .AddSingleton(new AIConfig().Build())
-                .AddSingleton<AIService>()
-                .AddSingleton<IndexerService>();
+                .AddSingleton(config)
+                .AddSingleton<IndexerService>()
+                .AddSingleton<AIService>();
 
             return 0;
         }
 
-        public int OnUnload()
+        public int OnUnload(ILogger logger)
         {
             return 0;
+        }
+        
+        private static bool LoadConfig(ILogger logger, out DuplicateDetectorConfig config)
+        {
+            var configPath = Path.Combine(PluginHelper.Directory(), "config.json");
+            if (!File.Exists(configPath))
+            {
+                logger.Log(LogSeverity.Warning,
+                    "Config doesn't exist! The file has been created, please edit all fields to be correct. Exiting.");
+                new DuplicateDetectorConfig().Write(configPath);
+                config = null;
+                return false;
+            }
+
+            config = new DuplicateDetectorConfig(logger, configPath);
+            return config.VerifyConfig();
         }
     }
 }
