@@ -1,14 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Dapper;
 using Discord;
 using Discord.WebSocket;
 using DuplicateDetector.Templates;
-using MySql.Data.MySqlClient;
 using Newtonsoft.Json;
 using UNObot.Plugins;
 using UNObot.Plugins.Helpers;
@@ -19,12 +17,12 @@ namespace DuplicateDetector.Services
     {
         private readonly ILogger _logger;
         private readonly DiscordSocketClient _client;
-        private readonly IAIConfig _config;
+        private readonly DuplicateDetectorConfig _config;
 
         private readonly string _cacheDir;
         private readonly string _imageDir;
         
-        public IndexerService(ILogger logger, DiscordSocketClient client, IAIConfig config)
+        public IndexerService(ILogger logger, DiscordSocketClient client, DuplicateDetectorConfig config)
         {
             _logger = logger;
             _client = client;
@@ -54,24 +52,34 @@ namespace DuplicateDetector.Services
         {
             Task.Run(async () =>
             {
-                await using var db = new MySqlConnection(_config.DBConnStr);
+                await using var db = _config.GetConnection();
 
                 foreach (var attachment in message.Attachments)
                 {
+                    Console.WriteLine("Starting write");
+                    try
+                    {
 #pragma warning disable 4014
-                    db.ExecuteAsync(
-                        "INSERT INTO DuplicateDetector.Images (channel, message, author, url, proxy_url, spoiler, posted) VALUES (@Channel, @Message, @Author, @URL, @Proxy_URL, @Spoiler, @Posted)",
-                        new 
-                        {
-                            Channel = message.Channel.Id,
-                            Author = message.Author.Id,
-                            Message = message.Id,
-                            URL = attachment.Url,
-                            Proxy_URL = attachment.ProxyUrl,
-                            Spoiler = attachment.IsSpoiler(),
-                            Posted = message.Timestamp.UtcDateTime
-                        });
+                        db.ExecuteAsync(
 #pragma warning restore 4014
+                            _config.ConvertSql(
+                                "INSERT INTO DuplicateDetector.Images (channel, message, author, url, proxy_url, spoiler, posted) VALUES (@Channel, @Message, @Author, @URL, @Proxy_URL, @Spoiler, @Posted)"),
+                            new
+                            {
+                                Channel = Convert.ToDecimal(message.Channel.Id),
+                                Author = Convert.ToDecimal(message.Author.Id),
+                                Message = Convert.ToDecimal(message.Id),
+                                URL = attachment.Url,
+                                Proxy_URL = attachment.ProxyUrl,
+                                Spoiler = attachment.IsSpoiler(),
+                                Posted = message.Timestamp.UtcDateTime
+                            });
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex);
+                    }
+
                 }
             });
         }
@@ -80,13 +88,11 @@ namespace DuplicateDetector.Services
         {
             try
             {
-                await using var db = new MySqlConnection(_config.DBConnStr);
+                await using var db = _config.GetConnection();
                 
-                var output = await db.QueryAsync<bool>(
-                    "SELECT autolog FROM DuplicateDetector.Channels WHERE channel = @Channel",
-                    new {Channel = channel});
-                
-                return output.SingleOrDefault();
+                return await db.ExecuteScalarAsync<bool>(
+                    _config.ConvertSql("SELECT autolog FROM DuplicateDetector.Channels WHERE channel = @Channel"),
+                    new { Channel = Convert.ToDecimal(channel) });
             }
             catch (Exception e)
             {
@@ -102,10 +108,11 @@ namespace DuplicateDetector.Services
                 var counter = 0;
                 var path = Path.Combine(_cacheDir, $"url_list-{channel.Id}.json");
                 await using var sw = new StreamWriter(path);
-                await using var db = new MySqlConnection(_config.DBConnStr);
+                await using var db = _config.GetConnection();
                 _ = db.ExecuteAsync(
-                    "INSERT INTO DuplicateDetector.Channels (channel, server, autolog) VALUES (@Channel, @Server, 1) ON DUPLICATE KEY UPDATE autolog = 1",
-                    new { Channel = channel.Id, Server = channel.Guild.Id });
+                    _config.ConvertSql(
+                        "INSERT INTO DuplicateDetector.Channels (channel, server, autolog) VALUES (@Channel, @Server, 1) ON DUPLICATE KEY UPDATE autolog = 1"
+                        ),new { Channel = Convert.ToDecimal(channel.Id), Server = Convert.ToDecimal(channel.Guild.Id) });
                 var writeMessages = new List<ImageMessage>();
                 _logger.Log(LogSeverity.Debug, "Starting indexer.");
                 await foreach (var messages in channel.GetMessagesAsync(100000))
@@ -148,20 +155,32 @@ namespace DuplicateDetector.Services
         public async Task Index(ulong guild, ulong channel)
             => await Index(_client.GetGuild(guild).GetTextChannel(channel));
 
+        private class Image
+        {
+            public Image(int id, string url)
+            {
+                ID = id;
+                URL = url;
+            }
+
+            public int ID { get; }
+            public string URL { get; }
+        }
+        
         public async Task Download()
         {
-            const string command = "SELECT id, url, proxy_url FROM DuplicateDetector.Images";
-            await using var connection = new MySqlConnection(_config.DBConnStr);
+            const string command = "SELECT id, url FROM DuplicateDetector.Images";
+            await using var db = _config.GetConnection();
             try
             {
-                connection.Open();
-                await using var mySQLCmd = new MySqlCommand(command, connection);
-                await using var reader = await mySQLCmd.ExecuteReaderAsync();
-                while (await reader.ReadAsync())
+                await using var query = await db.ExecuteReaderAsync(_config.ConvertSql(command));
+                var rowParser = query.GetRowParser<Image>();
+                
+                while (await query.ReadAsync())
                 {
-                    var url = reader.GetString(1) ?? reader.GetString(2);
+                    var data = rowParser(query);
 #pragma warning disable 4014
-                    DownloadImage(url, reader.GetInt32(0));
+                    DownloadImage(data.URL, data.ID);
 #pragma warning restore 4014
                 }
             }
@@ -169,10 +188,6 @@ namespace DuplicateDetector.Services
             {
                 _logger.Log(LogSeverity.Error, "Downloader failed!", e);
                 throw;
-            }
-            finally
-            {
-                await connection.CloseAsync();
             }
         }
         
