@@ -12,7 +12,7 @@ namespace UNObot.Core.Modules
         public string Version { get; private set; }
         public IServiceCollection Services { get; private set; }
 
-        public int OnLoad()
+        public int OnLoad(ILogger logger)
         {
             Name = "UNObot-Core";
             Description = "The core UNO commands the bot is not known for.";
@@ -22,15 +22,14 @@ namespace UNObot.Core.Modules
             Services = new ServiceCollection()
                 .AddSingleton<AFKTimerService>()
                 .AddSingleton<EmbedService>()
-                .AddSingleton<InputHandlerService>()
                 .AddSingleton<QueueHandlerService>()
-                .AddSingleton<UNODatabaseService>()
+                .AddSingleton<DatabaseService>()
                 .AddSingleton<UNOPlayCardService>();
             
             return 0;
         }
 
-        public int OnUnload()
+        public int OnUnload(ILogger logger)
         {
             return 0;
         }
