@@ -4,27 +4,27 @@ using System.Text;
 using System.Threading.Tasks;
 using Discord;
 using Discord.WebSocket;
-using Microsoft.Extensions.Configuration;
 using UNObot.MusicBot.MusicCore;
+using UNObot.Plugins;
 using UNObot.Plugins.TerminalCore;
 using YoutubeExplode.Playlists;
 
 namespace UNObot.MusicBot.Services
 {
-    internal class EmbedService
+    public class EmbedService
     {
         private readonly YoutubeService _youtube;
-        private readonly IConfiguration _config;
+        private readonly IUNObotConfig _config;
         private readonly DiscordSocketClient _client;
         
-        public EmbedService(YoutubeService youtube, IConfiguration config, DiscordSocketClient client)
+        public EmbedService(YoutubeService youtube, IUNObotConfig config, DiscordSocketClient client)
         {
             _youtube = youtube;
             _config = config;
             _client = client;
         }
         
-        internal Tuple<Embed, Tuple<string, string, string>> DisplayAddSong(ulong userId, ulong serverId,
+        public Tuple<Embed, Tuple<string, string, string>> DisplayAddSong(ulong userId, ulong serverId,
             string songUrl, Tuple<string, string, string> information)
         {
             var server = _client.GetGuild(serverId).Name;
@@ -39,7 +39,7 @@ namespace UNObot.MusicBot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithThumbnailUrl(information.Item3)
@@ -55,7 +55,7 @@ namespace UNObot.MusicBot.Services
             return new Tuple<Embed, Tuple<string, string, string>>(embed, information);
         }
 
-        internal Embed DisplayNowPlaying(Song song, string currentDuration)
+        public Embed DisplayNowPlaying(Song song, string currentDuration)
         {
             var username = _client.GetUser(song.RequestedBy).Username;
             var servername = _client.GetGuild(song.RequestedGuild).Name;
@@ -69,7 +69,7 @@ namespace UNObot.MusicBot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithThumbnailUrl(song.ThumbnailUrl)
@@ -85,7 +85,7 @@ namespace UNObot.MusicBot.Services
             return builder.Build();
         }
 
-        internal async Task<Tuple<Embed, Playlist>> DisplayPlaylist(ulong userId, ulong serverId, string songUrl)
+        public async Task<Tuple<Embed, Playlist>> DisplayPlaylist(ulong userId, ulong serverId, string songUrl)
         {
             var username = _client.GetUser(userId).Username;
             var servername = _client.GetGuild(serverId).Name;
@@ -101,7 +101,7 @@ namespace UNObot.MusicBot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithThumbnailUrl(thumbnail)
@@ -117,7 +117,7 @@ namespace UNObot.MusicBot.Services
             return new Tuple<Embed, Playlist>(builder.Build(), playlist);
         }
 
-        internal Tuple<Embed, int> DisplaySongList(Song nowPlaying, List<Song> songs, int page)
+        public Tuple<Embed, int> DisplaySongList(Song nowPlaying, List<Song> songs, int page)
         {
             var containers = new List<StringBuilder>();
             var r = ThreadSafeRandom.ThisThreadsRandom;
@@ -169,7 +169,7 @@ namespace UNObot.MusicBot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithAuthor(author =>
