@@ -1,8 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using System.Data.Common;
 using System.Threading.Tasks;
+using Dapper;
 using Discord;
-using Microsoft.Extensions.Configuration;
-using MySql.Data.MySqlClient;
 using UNObot.Plugins;
 using UNObot.Plugins.Helpers;
 
@@ -11,39 +10,29 @@ namespace UNObot.ServerQuery.Services
     public class DatabaseService
     {
         private readonly ILogger _logger;
-
-        internal string ConnString { get; }
+        private readonly IUNObotConfig _config;
         
-        public DatabaseService(ILogger logger, IConfiguration config)
+        public DatabaseService(ILogger logger, IUNObotConfig config)
         {
             _logger = logger;
-            ConnString = config.GetConnectionString();
+            _config = config;
         }
         
-        internal async Task<string> GetMinecraftUser(ulong user)
+        public async Task<string> GetMinecraftUser(ulong user)
         {
-            string username = null;
-            var parameters = new List<MySqlParameter>();
+            await using var db = _config.GetConnection();
 
-            const string commandText = "SELECT minecraftUsername FROM UNObot.Players WHERE userid = ?";
+            const string commandText = "SELECT minecraftUsername FROM UNObot.Players WHERE userid = @UserID";
 
-            var p1 = new MySqlParameter
-            {
-                Value = user
-            };
-            parameters.Add(p1);
             try
             {
-                var result = await MySqlHelper.ExecuteScalarAsync(ConnString, commandText, parameters.ToArray());
-                if (result.HasDBValue())
-                    username = (string) result;
+                return await db.ExecuteScalarAsync<string>(_config.ConvertSql(commandText), new { UserID = user } );
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+                return null;
             }
-
-            return username;
         }
     }
 }
