@@ -12,18 +12,18 @@ namespace UNObot.Modules
     {
         private readonly ILogger _logger;
         
-        internal FunCommands(ILogger logger)
+        public FunCommands(ILogger logger)
         {
             _logger = logger;
         }
 
         [Command("gulag", RunMode = RunMode.Async)]
         [Help(new[] {".gulag"}, "Blyat.", false, "UNObot 1.4")]
-        internal async Task Gulag() => await Gulag("" + Context.User.Id);
+        public async Task Gulag() => await Gulag("" + Context.User.Id);
 
         [Command("gulag", RunMode = RunMode.Async)]
         [Help(new[] {".gulag (user)"}, "Blyat.", false, "UNObot 1.4")]
-        internal async Task Gulag(string user)
+        public async Task Gulag(string user)
         {
             //extraclean
             user = user.Trim(' ', '<', '>', '!', '@');
@@ -32,7 +32,7 @@ namespace UNObot.Modules
 
         [Command("nepnep", RunMode = RunMode.Async)]
         [Help(new[] {".nepnep"}, "Wait, how did this command get in here?", false, "UNObot 1.4")]
-        internal async Task Nep()
+        public async Task Nep()
         {
             await ReplyAsync("You got me there at \"nep\".");
         }
@@ -40,7 +40,7 @@ namespace UNObot.Modules
         [Command("ugay", RunMode = RunMode.Async)]
         [Alias("u gay", "you gay", "you're gay")]
         [Help(new[] {".ugay"}, "That's not very nice. >:[", false, "UNObot 0.1")]
-        internal async Task Ugay()
+        public async Task Ugay()
         {
             await ReplyAsync(
                 $"<@{Context.User.Id}> no u\n");
@@ -49,7 +49,7 @@ namespace UNObot.Modules
         [Command("no u", RunMode = RunMode.Async)]
         [Alias("nou")]
         [Help(new[] {".no u"}, "Fite me m8", false, "UNObot 1.0")]
-        internal async Task Easteregg1()
+        public async Task Easteregg1()
         {
             await ReplyAsync($"I claim that <@{Context.User.Id}> is triple gay. Say \"No U\" again, uh...");
         }
@@ -57,28 +57,28 @@ namespace UNObot.Modules
         [Command("upupdowndownleftrightleftrightbastart", RunMode = RunMode.Async)]
         [Help(new[] {".upupdowndownleftrightleftrightbastart"}, "Wow, an ancient easter egg. It's still ancient.",
             false, "UNObot 1.4")]
-        internal async Task OldEasterEgg()
+        public async Task OldEasterEgg()
         {
             await ReplyAsync("...Did you seriously think that would work?");
         }
 
         [Command("moltthink", RunMode = RunMode.Async)]
         [Help(new[] {".moltthink"}, "Think like Molt.", true, "UNObot 3.0 Beta 1")]
-        internal async Task MoltThink()
+        public async Task MoltThink()
         {
             await ReplyAsync("<:moltthink:471842854591791104>");
         }
 
         [Command("moltthinkreact", RunMode = RunMode.Async)]
         [Help(new[] {".moltthinkreact"}, "React by thinking as Molt.", true, "UNObot 3.0 Beta 1")]
-        internal async Task MoltThinkReact()
+        public async Task MoltThinkReact()
         {
             await MoltThinkReact(1);
         }
 
         [Command("moltthinkreact", RunMode = RunMode.Async)]
         [Help(new[] {".moltthinkreact (number of messages)"}, "React by thinking as Molt.", true, "UNObot 3.0 Beta 1")]
-        internal async Task MoltThinkReact(int numMessages)
+        public async Task MoltThinkReact(int numMessages)
         {
             IEmote emote = await Context.Client.GetGuild(420005591155605535).GetEmoteAsync(471842854591791104);
             await BaseReact(numMessages, emote);
@@ -86,21 +86,21 @@ namespace UNObot.Modules
 
         [Command("oof", RunMode = RunMode.Async)]
         [Help(new[] {".oof"}, "Oof.", true, "UNObot 3.0 Beta 1")]
-        internal async Task Oof()
+        public async Task Oof()
         {
             await ReplyAsync("<:oof:559961296418635776>");
         }
 
         [Command("oofreact", RunMode = RunMode.Async)]
         [Help(new[] {".oofreact"}, "Damn, oof.", true, "UNObot 3.0 Beta 1")]
-        internal async Task OofReact()
+        public async Task OofReact()
         {
             await OofReact(1);
         }
 
         [Command("oofreact", RunMode = RunMode.Async)]
         [Help(new[] {".oofreact (number of messages)"}, "Damn, oof.", true, "UNObot 3.0 Beta 1")]
-        internal async Task OofReact(int numMessages)
+        public async Task OofReact(int numMessages)
         {
             IEmote emote = await Context.Client.GetGuild(420005591155605535).GetEmoteAsync(559961296418635776);
             await BaseReact(numMessages, emote);
@@ -109,14 +109,14 @@ namespace UNObot.Modules
         [Command("calculateemote", RunMode = RunMode.Async)]
         [RequireOwner]
         [DisableDMs]
-        internal async Task CalculateEmote([Remainder] string input)
+        public async Task CalculateEmote([Remainder] string input)
         {
             await ReplyAsync($"Server: ``{Context.Guild.Id}`` emoteID: ``{input}``").ConfigureAwait(false);
         }
 
         [Command("emote", RunMode = RunMode.Async)]
         [RequireOwner]
-        internal async Task Emote(ulong server, ulong emoteId)
+        public async Task Emote(ulong server, ulong emoteId)
         {
             try
             {
@@ -131,7 +131,7 @@ namespace UNObot.Modules
 
         [Command("emotereact", RunMode = RunMode.Async)]
         [RequireOwner]
-        internal async Task EmoteReact(ulong server, ulong emoteId, int numMessages)
+        public async Task EmoteReact(ulong server, ulong emoteId, int numMessages)
         {
             try
             {
@@ -144,7 +144,7 @@ namespace UNObot.Modules
             }
         }
 
-        internal async Task BaseReact(int numMessages, IEmote emote)
+        public async Task BaseReact(int numMessages, IEmote emote)
         {
             var messages = await Context.Channel.GetMessagesAsync(numMessages + 1).FlattenAsync();
             var message = messages.Last();
