@@ -3,19 +3,31 @@ using System.Linq;
 using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
+using UNObot.Plugins;
 using UNObot.Plugins.Attributes;
 using UNObot.Plugins.Helpers;
+using UNObot.Plugins.Settings;
 using UNObot.Services;
 
 namespace UNObot.Modules
 {
     public class SettingsCommands : ModuleBase<SocketCommandContext>
     {
+        private readonly IUNObotConfig _config;
         private readonly DatabaseService _db;
+        private readonly EmbedDisplayService _embed;
         
-        internal SettingsCommands(DatabaseService db)
+        public SettingsCommands(IUNObotConfig config, DatabaseService db, EmbedDisplayService embed)
         {
+            _config = config;
             _db = db;
+            _embed = embed;
+            
+            var settings = new Setting("General Settings");
+            settings.UpdateSetting("Prefix", new CodeBlock("."));
+            settings.UpdateSetting("Enforce Channels", new Boolean(false));
+            settings.UpdateSetting("Channels Enforced", new ChannelIDList());
+            SettingsManager.RegisterSettings("UNObot", settings);
         }
         
         [Command("setdefaultchannel", RunMode = RunMode.Async)]
@@ -24,7 +36,7 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".setdefaultchannel"}, "Set the default channel for UNObot to chat in. Managers only.", true,
             "UNObot 2.0")]
-        internal async Task SetDefaultChannel()
+        public async Task SetDefaultChannel()
         {
             await ReplyAsync($":white_check_mark: Set default UNO channel to #{Context.Channel.Name}.");
             await _db.SetDefaultChannel(Context.Guild.Id, Context.Channel.Id);
@@ -42,12 +54,12 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".removedefaultchannel"}, "Remove the default channel for UNObot to chat in. Managers only.", true,
             "UNObot 2.0")]
-        internal async Task RemoveDefaultChannel()
+        public async Task RemoveDefaultChannel()
         {
             await ReplyAsync(":white_check_mark: Removed default UNO channel, assuming there was one.");
-            if (!await DatabaseExtensions.HasDefaultChannel(_db.ConnString, Context.Guild.Id))
+            if (!await DatabaseExtensions.HasDefaultChannel(_config, Context.Guild.Id))
             {
-                var channel = await DatabaseExtensions.GetDefaultChannel(_db.ConnString, Context.Guild.Id);
+                var channel = await DatabaseExtensions.GetDefaultChannel(_config, Context.Guild.Id);
                 //remove default channel
                 var currentChannels = await _db.GetAllowedChannels(Context.Guild.Id);
                 currentChannels.Remove(channel);
@@ -65,7 +77,7 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".enforcechannels"},
             "Only allow UNObot to recieve commands from enforced channels. Managers only.", true, "UNObot 2.0")]
-        internal async Task EnforceChannel()
+        public async Task EnforceChannel()
         {
             //start check (make sure all channels exist at time of enforcing)
             var allowedChannels = await _db.GetAllowedChannels(Context.Guild.Id);
@@ -88,7 +100,7 @@ namespace UNObot.Modules
                 return;
             }
 
-            if (!await DatabaseExtensions.HasDefaultChannel(_db.ConnString, Context.Guild.Id))
+            if (!await DatabaseExtensions.HasDefaultChannel(_config, Context.Guild.Id))
             {
                 await Context.Channel.SendMessageAsync(
                     "Error: Cannot enable enforcechannels if there is no default channel!");
@@ -109,13 +121,13 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".addallowedchannel"}, "Allow the current channel to accept commands. Managers only.", true,
             "UNObot 2.0")]
-        internal async Task AddAllowedChannel()
+        public async Task AddAllowedChannel()
         {
-            if (!await DatabaseExtensions.HasDefaultChannel(_db.ConnString, Context.Guild.Id))
+            if (!await DatabaseExtensions.HasDefaultChannel(_config, Context.Guild.Id))
             {
                 await ReplyAsync("Error: You need to set a default channel first.");
             }
-            else if (await DatabaseExtensions.GetDefaultChannel(_db.ConnString, Context.Guild.Id) == Context.Channel.Id)
+            else if (await DatabaseExtensions.GetDefaultChannel(_config, Context.Guild.Id) == Context.Channel.Id)
             {
                 await ReplyAsync(
                     "The default UNO channel has been set to this already; there is no need to add this as a default channel.");
@@ -139,7 +151,7 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".listallowedchannels"},
             "See all channels that UNObot can accept commands if enforced mode was on.", true, "UNObot 2.0")]
-        internal async Task ListAllowedChannels()
+        public async Task ListAllowedChannels()
         {
             var allowedChannels = await _db.GetAllowedChannels(Context.Guild.Id);
             //start check
@@ -175,7 +187,7 @@ namespace UNObot.Modules
         [DisableDMs]
         [Help(new[] {".removeallowedchannel"},
             "Remove a channel that UNObot previously was allowed to accept commands from.", true, "UNObot 2.0")]
-        internal async Task RemoveAllowedChannel()
+        public async Task RemoveAllowedChannel()
         {
             //start check
             var allowedChannels = await _db.GetAllowedChannels(Context.Guild.Id);
@@ -195,12 +207,23 @@ namespace UNObot.Modules
             {
                 allowedChannels.Remove(Context.Channel.Id);
                 await _db.SetAllowedChannels(Context.Guild.Id, allowedChannels);
-                await ReplyAsync($"Removed <#{Context.Channel.Id}> from the allowed channels!");
+                await ReplyAsync($":white_check_mark: Removed <#{Context.Channel.Id}> from the allowed channels!");
             }
             else
             {
-                await ReplyAsync("This channel was never an allowed channel.");
+                await ReplyAsync(":no_entry: This channel was never an allowed channel.");
             }
+        }
+
+        [Command("settings", RunMode = RunMode.Async)]
+        [RequireUserPermission(GuildPermission.ManageChannels)]
+        [DisableDMs]
+        [Help(new[] {".settings"},
+            "Access configurable settings for UNObot.", true, "UNObot 4.3")]
+        public async Task ViewSettings()
+        {
+            var manager = await _db.GetSettings(Context.Guild.Id);
+            await ReplyAsync("", embed: _embed.SettingsEmbed(manager.CurrentSettings.Values));
         }
     }
 }
