@@ -1,150 +1,104 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Common;
 using System.Threading.Tasks;
+using Dapper;
 using Discord;
-using Microsoft.Extensions.Configuration;
-using MySql.Data.MySqlClient;
 using Newtonsoft.Json;
 using UNObot.Plugins;
 using UNObot.Plugins.Helpers;
+using UNObot.Plugins.Settings;
 
 namespace UNObot.Services
 {
-    internal class DatabaseService
+    public class DatabaseService
     {
+        private readonly IUNObotConfig _config;
         private readonly ILogger _logger;
 
-        internal string ConnString { get; }
-
-        public DatabaseService(ILogger logger, IConfiguration config)
+        public DatabaseService(IUNObotConfig config, ILogger logger)
         {
+            _config = config;
             _logger = logger;
-            ConnString = config.GetConnectionString();
         }
 
-        internal async Task SetDefaultChannel(ulong server, ulong channel)
+        public async Task SetDefaultChannel(ulong server, ulong channel)
         {
-            var parameters = new List<MySqlParameter>();
+            const string commandText = "UPDATE UNObot.Games SET playChannel = @Channel WHERE server = @Server";
 
-            const string commandText = "UPDATE Games SET playChannel = ? WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = channel
-            };
-            parameters.Add(p1);
-            var p2 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p2);
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new { Channel = Convert.ToDecimal(channel), Server = Convert.ToDecimal(server) });
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task SetHasDefaultChannel(ulong server, bool hasDefault)
+        public async Task SetHasDefaultChannel(ulong server, bool hasDefault)
         {
-            var parameters = new List<MySqlParameter>();
+            const string commandText = "UPDATE UNObot.Games SET hasDefaultChannel = @HasDefaultChannel WHERE server = @Server";
 
-            const string commandText = "UPDATE Games SET hasDefaultChannel = ? WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = hasDefault
-            };
-            parameters.Add(p1);
-            var p2 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p2);
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {HasDefaultChannel = hasDefault, Server = Convert.ToDecimal(server)});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task<bool> ChannelEnforced(ulong server)
+        public async Task<bool> ChannelEnforced(ulong server)
         {
-            var yesOrNo = false;
-            var parameters = new List<MySqlParameter>();
+            const string commandText = "SELECT enforceChannel FROM UNObot.Games WHERE server = @Server";
+            var result = false;
 
-            const string commandText = "SELECT enforceChannel FROM UNObot.Games WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p1);
+            await using var db = _config.GetConnection();
             try
             {
-                var result = await MySqlHelper.ExecuteScalarAsync(ConnString, commandText, parameters.ToArray());
-                if (result.HasDBValue())
-                    yesOrNo = (byte) result == 1;
+                result = await db.ExecuteScalarAsync<bool>(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server)});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
 
-            return yesOrNo;
+            return result;
         }
 
-        internal async Task SetEnforceChannel(ulong server, bool enforce)
+        public async Task SetEnforceChannel(ulong server, bool enforce)
         {
-            var parameters = new List<MySqlParameter>();
+            const string commandText = "UPDATE UNObot.Games SET enforceChannel = @Enforce WHERE server = @Server";
 
-            const string commandText = "UPDATE Games SET enforceChannel = ? WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = enforce
-            };
-            parameters.Add(p1);
-            var p2 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p2);
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {Enforce = enforce, Server = Convert.ToDecimal(server)});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task<List<ulong>> GetAllowedChannels(ulong server)
+        public async Task<List<ulong>> GetAllowedChannels(ulong server)
         {
+            const string commandText = "SELECT allowedChannels FROM UNObot.Games WHERE server = @Server";
+
             var allowedChannels = new List<ulong>();
-            var parameters = new List<MySqlParameter>();
 
-            const string commandText = "SELECT allowedChannels FROM UNObot.Games WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p1);
+            await using var db = _config.GetConnection();
             try
             {
-                var result = await MySqlHelper.ExecuteScalarAsync(ConnString, commandText, parameters.ToArray());
-                if(result.HasDBValue())
+                var result = await db.ExecuteScalarAsync(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server)});
+                if (result.HasDBValue())
                     allowedChannels = JsonConvert.DeserializeObject<List<ulong>>((string) result);
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
@@ -152,166 +106,177 @@ namespace UNObot.Services
             return allowedChannels;
         }
 
-        internal async Task SetAllowedChannels(ulong server, List<ulong> allowedChannels)
+        public async Task SetAllowedChannels(ulong server, List<ulong> allowedChannels)
         {
-            var parameters = new List<MySqlParameter>();
+            const string commandText = "UPDATE UNObot.Games SET allowedChannels = @AllowedChannels WHERE server = @Server";
+            var json = JsonConvert.SerializeObject(allowedChannels);
 
-            const string commandText = "UPDATE Games SET allowedChannels = ? WHERE server = ?";
-
-            var p1 = new MySqlParameter
-            {
-                Value = JsonConvert.SerializeObject(allowedChannels)
-            };
-            parameters.Add(p1);
-            var p2 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p2);
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server), AllowedChannels = json});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task AddUser(ulong id, string username)
-        {
-            var parameters = new List<MySqlParameter>();
-
-            const string commandText =
-                "INSERT INTO Players (userid, username) VALUES(?, ?) ON DUPLICATE KEY UPDATE username = ?";
-            var p1 = new MySqlParameter
-            {
-                Value = id
-            };
-            parameters.Add(p1);
-            var p2 = new MySqlParameter
-            {
-                Value = username
-            };
-            parameters.Add(p2);
-            var p3 = new MySqlParameter
-            {
-                Value = username
-            };
-            parameters.Add(p3);
-            try
-            {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
-            }
-            catch (MySqlException ex)
-            {
-                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
-            }
-        }
-        
-        internal async Task AddGame(ulong server)
-        {
-            const string commandText = "INSERT IGNORE INTO Games (server) VALUES(?)";
-            var parameters = new List<MySqlParameter>();
-            var p1 = new MySqlParameter
-            {
-                Value = server
-            };
-            parameters.Add(p1);
-
-            try
-            {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
-            }
-            catch (MySqlException ex)
-            {
-                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
-            }
-        }
-        
-        internal async Task AddWebhook(string key, ulong guild, ulong channel, string type = "bitbucket")
+        public async Task RegisterUser(ulong id, string username)
         {
             const string commandText =
-                "INSERT INTO UNObot.Webhooks (webhookKey, channel, guild, type) VALUES (?, ?, ?, ?)";
+                "INSERT INTO UNObot.Players (userid, username) VALUES(@UserID, @Username) ON DUPLICATE KEY UPDATE username = @Username";
 
-            var p1 = new MySqlParameter
-            {
-                Value = key.Substring(0, 50)
-            };
-            var p2 = new MySqlParameter
-            {
-                Value = channel
-            };
-            var p3 = new MySqlParameter
-            {
-                Value = guild
-            };
-            var p4 = new MySqlParameter
-            {
-                Value = type
-            };
-
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, p1, p2, p3, p4);
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {UserID = Convert.ToDecimal(id), Username = username});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task DeleteWebhook(string key)
+        public async Task RegisterServer(ulong server)
         {
-            const string commandText = "DELETE FROM UNObot.Webhooks WHERE webhookKey = ?";
-            var parameters = new List<MySqlParameter>();
-            var p1 = new MySqlParameter
-            {
-                Value = key.Substring(0, Math.Min(key.Length, 50))
-            };
-            parameters.Add(p1);
+            const string commandText = "INSERT IGNORE INTO UNObot.Games (server) VALUES(@Server)";
 
+            await using var db = _config.GetConnection();
             try
             {
-                await MySqlHelper.ExecuteNonQueryAsync(ConnString, commandText, parameters.ToArray());
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server)});
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
         }
 
-        internal async Task<(ulong Guild, byte Type)> GetWebhook(ulong channel, string key)
+        public async Task AddWebhook(string key, ulong guild, ulong channel, string type = "bitbucket")
         {
-            const string commandText = "SELECT guild, type FROM UNObot.Webhooks WHERE channel = ? AND webhookKey = ?";
+            const string commandText =
+                "INSERT INTO UNObot.Webhooks (webhookKey, channel, guild, type) VALUES (@Key, @Channel, @Guild, @Type)";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                await db.ExecuteAsync(_config.ConvertSql(commandText),
+                    new {Key = key[..50], Channel = Convert.ToDecimal(channel), Guild = Convert.ToDecimal(guild), Type = type});
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
+        }
+
+        public async Task DeleteWebhook(string key)
+        {
+            const string commandText = "DELETE FROM UNObot.Webhooks WHERE webhookKey = @Key";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {Key = key.Substring(0, Math.Min(key.Length, 50))});
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
+        }
+
+        public async Task<(ulong Guild, byte Type)> GetWebhook(ulong channel, string key)
+        {
+            const string commandText =
+                "SELECT guild, type FROM UNObot.Webhooks WHERE channel = @Channel AND webhookKey = @Key";
             ulong guild = 0;
             byte type = 0;
-            var parameters = new List<MySqlParameter>();
-            var p1 = new MySqlParameter
-            {
-                Value = channel
-            };
-            var p2 = new MySqlParameter
-            {
-                Value = key.Substring(0, Math.Min(key.Length, 50))
-            };
-            parameters.Add(p1);
-            parameters.Add(p2);
-            await using var dr = await MySqlHelper.ExecuteReaderAsync(ConnString, commandText, parameters.ToArray());
+
+            await using var db = _config.GetConnection();
             try
             {
-                while (dr.Read())
-                {
-                    guild = dr.GetUInt64(0);
-                    type = dr.GetByte(1);
-                }
+                var result = await db.QueryFirstOrDefaultAsync(_config.ConvertSql(commandText),
+                    new {Channel = Convert.ToDecimal(channel), Key = key[..Math.Min(key.Length, 50)]});
+                guild = Convert.ToUInt64(result.guild);
+                type = result.type;
             }
-            catch (MySqlException ex)
+            catch (DbException ex)
             {
                 _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
             }
 
             return (guild, type);
+        }
+
+        public async Task<string> GetPrefix(ulong server)
+        {
+            const string commandText = "SELECT commandPrefix FROM UNObot.Games WHERE server = @Server";
+            var prefix = ".";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                prefix = await db.ExecuteScalarAsync<string>(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server)});
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
+
+            return prefix;
+        }
+
+        public async Task SetPrefix(ulong server, string prefix)
+        {
+            const string commandText = "UPDATE UNObot.Games SET commandPrefix = @Prefix WHERE server = @Server";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new {Prefix = prefix, Server = Convert.ToDecimal(server)});
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
+        }
+
+        public async Task<SettingsManager> GetSettings(ulong server)
+        {
+            const string commandText = "SELECT settings FROM UNObot.Games WHERE server = @Server";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                var result = await db.ExecuteScalarAsync<string>(_config.ConvertSql(commandText), new {Server = Convert.ToDecimal(server)});
+                if (result.HasDBValue())
+                    return JsonConvert.DeserializeObject<SettingsManager>(result);
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
+            catch (JsonSerializationException)
+            {
+                // ¯\_(ツ)_/¯ Deal with it! (Recreate the settings manager)
+            }
+
+            return new SettingsManager();
+        }
+
+        public async Task SetSettings(ulong server, ulong settings)
+        {
+            const string commandText = "UPDATE UNObot.Games SET settings = @Settings WHERE server = @Server";
+
+            await using var db = _config.GetConnection();
+            try
+            {
+                await db.ExecuteAsync(_config.ConvertSql(commandText), new { Settings = settings, Server = Convert.ToDecimal(server)});
+            }
+            catch (DbException ex)
+            {
+                _logger.Log(LogSeverity.Error, "A MySQL error has occurred.", ex);
+            }
         }
     }
 }
