@@ -1,22 +1,26 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using Discord;
-using Microsoft.Extensions.Configuration;
 using Newtonsoft.Json.Linq;
+using UNObot.Plugins;
 using UNObot.Plugins.Helpers;
+using UNObot.Plugins.Settings;
 using UNObot.Plugins.TerminalCore;
 
 namespace UNObot.Services
 {
-    internal class EmbedDisplayService
+    public class EmbedDisplayService
     {
-        private readonly IConfiguration _config;
+        private readonly IUNObotConfig _config;
 
-        public EmbedDisplayService(IConfiguration config)
+        public EmbedDisplayService(IUNObotConfig config)
         {
             _config = config;
         }
 
-        internal Embed WebhookEmbed(WebhookListenerService.CommitInfo info)
+        public Embed WebhookEmbed(WebhookListenerService.CommitInfo info)
         {
             var random = ThreadSafeRandom.ThisThreadsRandom;
 
@@ -26,7 +30,7 @@ namespace UNObot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithAuthor(author =>
@@ -41,7 +45,7 @@ namespace UNObot.Services
             return builder.Build();
         }
 
-        internal Embed OctoprintEmbed(WebhookListenerService.OctoprintInfo info)
+        public Embed OctoprintEmbed(WebhookListenerService.OctoprintInfo info)
         {
             var random = ThreadSafeRandom.ThisThreadsRandom;
 
@@ -51,7 +55,7 @@ namespace UNObot.Services
                 .WithFooter(footer =>
                 {
                     footer
-                        .WithText($"UNObot {_config["version"]} - By DoggySazHi")
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
                         .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
                 })
                 .WithAuthor(author =>
@@ -78,6 +82,44 @@ namespace UNObot.Services
             if (bytesPrinted != null && bytesPrinted.Type != JTokenType.Null)
                 builder.AddField("Bytes Printed", (bytesPrinted.ToObject<float>() / 1000000.0).ToString("N2") + " MB",
                     true);
+            return builder.Build();
+        }
+
+        public Embed SettingsEmbed(IEnumerable<Setting> settings)
+        {
+            var random = ThreadSafeRandom.ThisThreadsRandom;
+
+            var builder = new EmbedBuilder()
+                .WithColor(new Color(random.Next(0, 256), random.Next(0, 256), random.Next(0, 256)))
+                .WithTimestamp(DateTimeOffset.Now)
+                .WithFooter(footer =>
+                {
+                    footer
+                        .WithText($"UNObot {_config.Version} - By DoggySazHi")
+                        .WithIconUrl("https://williamle.com/unobot/doggysazhi.png");
+                })
+                .WithAuthor(author =>
+                {
+                    author
+                        .WithName("UNObot Settings")
+                        .WithIconUrl("https://williamle.com/unobot/unobot.png");
+                });
+            foreach (var group in settings)
+            {
+                var titleLength = group.Relation.Keys.Max(o => o.Length) + 1;
+                var sb = new StringBuilder();
+                foreach (var key in group.Relation.Keys)
+                {
+                    sb.Append($"`{key.PadRight(titleLength)}|` ");
+                    var obj = group.GetSetting(key);
+                    if (obj == null)
+                        sb.Append("*None set*\n");
+                    else
+                        sb.Append(obj.Display).Append('\n');
+                }
+                builder.AddField(group.Category, sb.ToString());
+            }
+
             return builder.Build();
         }
     }
