@@ -4,10 +4,10 @@ using Newtonsoft.Json;
 
 namespace UNObot.Templates
 {
-    internal class Command
+    public class Command
     {
         [JsonConstructor]
-        internal Command(string commandName, List<string> aliases, List<string> usages, string help, bool active,
+        public Command(string commandName, List<string> aliases, List<string> usages, string help, bool active,
             string version)
         {
             CommandName = commandName;
@@ -19,20 +19,20 @@ namespace UNObot.Templates
         }
 
         [JsonConstructor]
-        internal Command(string commandName, List<string> aliases, List<string> usages, string help, bool active,
+        public Command(string commandName, List<string> aliases, List<string> usages, string help, bool active,
             string version, bool disableDMs) : this(commandName, aliases, usages, help, active, version)
         {
             DisableDMs = disableDMs;
         }
 
-        internal string CommandName { get; set; }
-        internal List<string> Usages { get; set; }
-        internal List<string> Aliases { get; set; }
-        internal string Help { get; set; }
-        internal bool Active { get; set; }
-        internal string Version { get; set; }
-        internal bool DisableDMs { get; set; }
-        internal IServiceProvider Services { get; set; }
-        internal bool Original { get; set; }
+        public string CommandName { get; set; }
+        public List<string> Usages { get; set; }
+        public List<string> Aliases { get; set; }
+        public string Help { get; set; }
+        public bool Active { get; set; }
+        public string Version { get; set; }
+        public bool DisableDMs { get; set; }
+        public IServiceProvider Services { get; set; }
+        public bool Original { get; set; }
     }
 }
